Added custom BandwidthSelect component,
Display an error if the bandwidth is higher then the lowest port
(SSP / MSP / ELAN) chosen by the customer. Updates every time a new
port is added, updated or deleted. It uses two new endpoint on the
server to fetch the port_speed fixed_input based on the product_id (SSP)
or the subscription_id (MSP).