New-Process page revamp
KlantSupport - and development - can now create new processes, modify
existing subscriptions and terminate subscriptions.

The page supports three predefined query parameters from KlantSupport:
product - the product guid which pre-selects the product for new
processes
organisation - the organisation CRM guid which is used to select the
subscriptions for both the modify and terminate
dienstfname - the CRM dienstafname which pre-fills the user input for
both the create and modify workflow

Only active and synced subscriptions can be modified and subscriptions
used as ports in existing non-terminated LP can also not be terminated.