Fix error loop
when showErrorDialog is true any backend connection error is also thrown outside the call
stack using setTimeout this gets picked up by the window.onerror function and is posted
tp the backend. If that results in an error you get a loop.

Fixed this by setting showErrorDialog to false for the call that pushes errors to the backend
Now you get only two errors, one original error and a failed push of that error