Fix selecting MSCs in multiservice port select
Because port_mode is not a resource type on MSC's it does not get send to
the frontend like with the other types. There was a workaround for that but I broke that when removing the code
that also send the tag and port_mode back as part of the form.
We now can reuse the getPortMode function to implement that properly so we don't have to modify the
response from the server..